backend/requirements.txt
alembic==1.8.1
aniso8601==6.0.0
click==8.1.3
colorama==0.4.5
flake8==5.0.4
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
mccabe==0.7.0
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
